Authorization

+++ app/Http/Controllers/PostController.php
@@ -15,7 +15,11 @@ class PostController extends Controller
     public function index() {
 
         // $posts = Post::get();
-        $posts = Post::with(['user', 'likes'])->paginate(20); // with(['user', 'likes']) ==> Eager loading (para reduzir e nº de querys). 'user' e 'likes' são funções de relacionamento do model Post
+        $posts = Post::latest()->with(['user', 'likes'])->paginate(20); 
+        
+        // with(['user', 'likes']) ==> Eager loading (para reduzir e nº de querys). 'user' e 'likes' são funções de relacionamento do model Post
+
+        // o orderBy('created_at', 'desc') pode ser substituído por latest() 
 
         return view('posts.index', [
             'posts' => $posts
@@ -46,6 +50,15 @@ class PostController extends Controller
 
     public function destroy(Post $post) {
         
+        /* essa instrução deve ser colocada no controller para que ninguém que não seja o autor do post consiga executá-la, mesmo modificando o caminho do action do formulário no html 
+        if (!$post->ownedBy(auth()->user())) {
+            // dd('Não');
+            return redirect('https://markhjorth.github.io/nedry/');
+        }
+        */
+
+        $this->authorize('delete', $post); // a função 'delete' está no PostPolicy
+
         $post->delete();
 
         return back();
